# Implement LZ4 transport compression