--- a/.gitignore
+++ b/.gitignore
@@ -2,3 +2,6 @@
 docs/build
 docs/Manifest.toml
 Manifest.toml
+
+# editor(s)
+.vscode