--- a/src/ClickHouse.jl
+++ b/src/ClickHouse.jl
@@ -9,6 +9,9 @@ include("Connect.jl")
 include("Query.jl")
 
 export Compression
+export COMPRESSION_NONE
+export COMPRESSION_LZ4
+export COMPRESSION_CHECKSUM_ONLY
 export ClickHouseSock
 export Block
 export select