--- a/src/Connect.jl
+++ b/src/Connect.jl
@@ -96,7 +96,7 @@ function connect(
     connection_timeout = DBMS_DEFAULT_CONNECT_TIMEOUT,
     max_insert_block_size = DBMS_DEFAULT_MAX_INSERT_BLOCK,
     send_buffer_size = DBMS_DEFAULT_BUFFER_SIZE,
-    compression::Union{Bool, AbstractString, UInt8, Compression} = false
+    compression::Compression = COMPRESSION_NONE
 )::ClickHouseSock
     sock = ClickHouseSock(
         nothing,
@@ -109,7 +109,7 @@ function connect(
             connection_timeout = connection_timeout,
             max_insert_block_size = max_insert_block_size,
             send_buffer_size = send_buffer_size,
-            compression = Compression(compression),
+            compression = compression,
         )
     )