--- a/src/tcp/Compression.jl
+++ b/src/tcp/Compression.jl
@@ -1,0 +1,33 @@
+
+@enum Compression::UInt8 begin
+    COMPRESSION_NONE = 0
+    COMPRESSION_DRY = 0x02
+    COMPRESSEION_LZ4 = 0x82
+end
+
+Compression(flag::Bool)::Compression = flag ? COMPRESSEION_LZ4 : COMPRESSION_NONE
+
+function Compression(name::String)::Compression
+    if lowercase(name) == "lz4"
+        return COMPRESSEION_LZ4
+    end
+    error("unkown compression mode: $(name)")
+end
+
+"""compress data according to the compression mode"""
+function compress(mode::Compression, data::Vector{UInt8})::Vector{UInt8}
+    return if mode == COMPRESSION_NONE || mode == COMPRESSION_DRY
+        data
+    elseif mode == COMPRESSION_LZ4
+        lz4_compress(data)
+    end
+end
+
+"""decompress data according to the compression mode"""
+function decompress(mode::Compression, data::Vector{UInt8})::Vector{UInt8}
+    return if mode == COMPRESSION_NONE || mode == COMPRESSION_DRY
+        data
+    elseif mode == COMPRESSION_LZ4
+        lz4_decompress(data)
+    end
+end