--- a/src/tcp/Compression.jl
+++ b/src/tcp/Compression.jl
@@ -2,14 +2,16 @@
 @enum Compression::UInt8 begin
     COMPRESSION_NONE = 0
     COMPRESSION_DRY = 0x02
-    COMPRESSEION_LZ4 = 0x82
+    COMPRESSION_LZ4 = 0x82
 end
 
-Compression(flag::Bool)::Compression = flag ? COMPRESSEION_LZ4 : COMPRESSION_NONE
+Compression(flag::Bool)::Compression = flag ? COMPRESSION_LZ4 : COMPRESSION_NONE
 
 function Compression(name::String)::Compression
     if lowercase(name) == "lz4"
-        return COMPRESSEION_LZ4
+        return COMPRESSION_LZ4
+    elseif  lowercase(name) == "dry"
+        return COMPRESSION_DRY
     end
     error("unkown compression mode: $(name)")
 end
@@ -19,6 +21,7 @@ function compress(mode::Compression, data::Vector{UInt8})::Vector{UInt8}
     return if mode == COMPRESSION_NONE || mode == COMPRESSION_DRY
         data
     elseif mode == COMPRESSION_LZ4
+        println("DECOMPRESSING :)")
         lz4_compress(data)
     end
 end
@@ -28,6 +31,11 @@ function decompress(mode::Compression, data::Vector{UInt8})::Vector{UInt8}
     return if mode == COMPRESSION_NONE || mode == COMPRESSION_DRY
         data
     elseif mode == COMPRESSION_LZ4
+        println("COMPRESSING :)")
         lz4_decompress(data)
     end
+end
+
+function chwrite(sock::ClickHouseSock, compression::ClickHouse.Compression)
+    chwrite(sock, UInt8(compression))
 end