PR for #8. Opening here in case there are interim comments, suggestions, or requirements. I have seen the @athre0z [branch](https://github.com/JuliaDatabases/ClickHouse.jl/tree/feature/transport-lz4) and will incorporate as necessary. Plans are to finish remaining work as time permits. 

- [x]  Develop a CityHash128 library
- [x] Test CityHash128 against ch hash values (only been hand tested)
- [x] Implement new `CompressedBlock` or alternative refactoring