@DylanModesitt 2022-10-02
If time permits, I'd like to solicit some feedback on this implementation. 

I can add more tests explicitly around reading/writing compressed blocks if desired. I stuck with @athre0z's sketch for `chread` and `chwrite` on the same `Block` structure instead of adding some more types & doing value dispatch on the compression method or something along those lines -- but happy to refactor as desired. 

___

As an aside, this is just one of the features that I wanted for my own use, but my current plans are to implement the [DBInterface](https://github.com/JuliaDatabases/DBInterface.jl) next. This will require something akin to #9 - [`clickhouse-driver`](https://github.com/mymarilyn/clickhouse-driver) seems to have a simple implementation of that which I plan to follow. Following that, I'll beef up the docs. 

I also think we should maybe tear out the explicit query building on insert (`"INSERT INTO $(table) VALUES"`) and instead overload `execute` to take an `iter` wherein the `ddl_query` would be passed from the user as `"INSERT INTO {{table}} VALUES"` which I believe is more safer, more flexible, and more intuitive for users. We can move these discussions to issues but just wanted to jot down some thoughts.