@DylanModesitt 2022-10-08
`city_hash_128` performance is now appropriate due to not allocating in `fetch64` and friends. Benchmark for a `523001` length string.

Before:

```
BenchmarkTools.Trial: 260 samples with 1 evaluation.
 Range (min … max):  18.340 ms … 22.842 ms  ┊ GC (min … max): 0.00% … 16.17%
 Time  (median):     18.621 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   19.274 ms ±  1.365 ms  ┊ GC (mean ± σ):  3.52% ±  6.18%

    ▂██                                                        
  ▂▄███▇▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▅▅▄▃▂ ▂
  18.3 ms         Histogram: frequency by time        22.2 ms <

 Memory estimate: 10.48 MiB, allocs estimate: 245180.
```

After:


```
BenchmarkTools.Trial: 6759 samples with 1 evaluation.
 Range (min … max):  684.250 μs …   3.343 ms  ┊ GC (min … max): 0.00% … 78.02%
 Time  (median):     717.375 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   738.717 μs ± 202.423 μs  ┊ GC (mean ± σ):  2.70% ±  7.25%

                ▁▅▇█▇▄▂▁                                         
  ▂▄▃▂▂▂▂▂▂▂▂▂▃▅████████▆▆▅▄▄▃▃▃▃▃▂▃▂▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▁▁▁▁▁▁▂▂ ▃
  684 μs           Histogram: frequency by time          795 μs <

 Memory estimate: 510.98 KiB, allocs estimate: 5.
```

___

The other suggested changes were also implemented (thanks)