@DylanModesitt 2022-10-16
> Thanks for the update and your patience -- my time has unfortunately been rather limited in the past few days and will likely continue to be so for the next two weeks. I'll try to carve out some time tomorrow for a final review pass.

No worries -- appreciate your thoughtful review! The comments you've made have been resolved. I will also see if I can squeeze a PR into `CodecLz4.jl` to fix the GC issue upstream and then later remove it here in a new PR. 

> In the mean time: I noticed that the DelimitedFiles package is being used in the tests while not being listed in Project.toml, causing the tests to fail unless manually installed up front. We'll either want to add this [as a test-dependency](https://pkgdocs.julialang.org/v1/creating-packages/#Test-specific-dependencies-in-Julia-1.2-and-above) or stop using this package; adding a general dependency just for the tests seems a bit overkill. I don't really have a preference here: feel free to pick what is easier for you.

Resolved - unnecessary dep.